# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RepoInsight API 文档",
        "title": "RepoInsight API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/repositories": {
            "get": {
                "description": "获取所有已分析的仓库列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "获取仓库列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/repositories/search": {
            "get": {
                "description": "根据关键词搜索仓库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "搜索仓库",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/repositories/{id}": {
            "get": {
                "description": "获取指定仓库的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "获取仓库详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Repository"
                        }
                    }
                }
            }
        },
        "/api/v1/status": {
            "get": {
                "description": "获取系统运行状态和统计信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取系统状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SystemStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.Repository": {
            "type": "object",
            "properties": {
                "ai_analysis": {
                    "description": "AI 分析后的介绍",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_branch": {
                    "description": "默认分支",
                    "type": "string"
                },
                "description": {
                    "description": "项目描述",
                    "type": "string"
                },
                "forks": {
                    "description": "分支数",
                    "type": "integer"
                },
                "full_name": {
                    "description": "项目全名，如 \"owner/repo\"",
                    "type": "string"
                },
                "has_downloads": {
                    "description": "是否启用下载",
                    "type": "boolean"
                },
                "has_issues": {
                    "description": "是否启用问题跟踪",
                    "type": "boolean"
                },
                "has_pages": {
                    "description": "是否启用 Pages",
                    "type": "boolean"
                },
                "has_projects": {
                    "description": "是否启用项目",
                    "type": "boolean"
                },
                "has_wiki": {
                    "description": "是否启用 Wiki",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_archived": {
                    "description": "是否已归档",
                    "type": "boolean"
                },
                "is_template": {
                    "description": "是否是模板仓库",
                    "type": "boolean"
                },
                "language": {
                    "description": "主要编程语言",
                    "type": "string"
                },
                "last_pushed_at": {
                    "description": "最后推送时间",
                    "type": "string"
                },
                "license": {
                    "description": "许可证",
                    "type": "string"
                },
                "name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "open_issues": {
                    "description": "开放问题数",
                    "type": "integer"
                },
                "owner": {
                    "description": "项目所有者",
                    "type": "string"
                },
                "readme": {
                    "description": "原始 README 内容",
                    "type": "string"
                },
                "size": {
                    "description": "仓库大小（KB）",
                    "type": "integer"
                },
                "stars": {
                    "description": "星标数",
                    "type": "integer"
                },
                "topics": {
                    "description": "项目标签，JSON 格式存储",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "GitHub 项目地址",
                    "type": "string"
                },
                "watchers": {
                    "description": "关注者数",
                    "type": "integer"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/api.Meta"
                }
            }
        },
        "api.SystemStatus": {
            "type": "object",
            "properties": {
                "last_updated": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_repositories": {
                    "type": "integer"
                }
            }
        }
    }
}